feat: Implement Employee Shift Management Module
- Added EmployeeShiftController for handling shift-related API requests.
- Created EmployeeShift entity to represent employee shifts in the database.
- Developed DTOs for creating and updating shifts (CreateShiftRequestDto, UpdateShiftRequestDto).
- Implemented response DTOs for shift details (EmployeeShiftResponseDto, EmployeeBasicDto, WorkShiftBasicDto).
- Introduced enums for shift status (EmployeeShiftStatus) and shift source (ShiftSource).
- Created EmployeeShiftMapper for converting between entities and DTOs.
- Established EmployeeShiftRepository for database operations on EmployeeShift entities.
- Developed EmployeeShiftService to encapsulate business logic for managing shifts.
- Added utility class EmployeeShiftValidator for shift-related validation rules.
- Implemented custom exceptions for handling specific error cases (e.g., EmployeeShiftNotFoundException, ShiftConflictException).

+++ src/main/java/com/dental/clinic/management/contact_history/dto/response/ContactHistoryResponse.java
@@ -1,6 +1,5 @@
 package com.dental.clinic.management.contact_history.dto.response;
 
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,6 +8,7 @@ import lombok.Setter;
 import java.time.LocalDateTime;
 
 import com.dental.clinic.management.contact_history.enums.ContactHistoryAction;
+
 /**
  * Response DTO for a contact interaction / history record.
  */
@@ -19,7 +19,7 @@ import com.dental.clinic.management.contact_history.enums.ContactHistoryAction;
 public class ContactHistoryResponse {
     private String historyId;
     private String contactId;
-    private Integer employeeId;
+    private Long employeeId;
     private ContactHistoryAction action;
     private String content;
     private LocalDateTime createdAt;
