+++ src/main/java/com/dental/clinic/management/customer_contact/repository/CustomerContactRepository.java
@@ -18,5 +18,5 @@ public interface CustomerContactRepository extends JpaRepository<CustomerContact
     long countByCreatedAtBetween(LocalDateTime from, LocalDateTime to);
 
     // new helper used for auto-assign + simple metrics
-    long countByAssignedToAndStatus(Integer assignedTo, CustomerContactStatus status);
+    long countByAssignedToAndStatus(Long assignedTo, CustomerContactStatus status);
 }
